🚚 auth와 관련된 헬퍼 함수를 파일로 분리
