+++ src/mocks/handlers/auth.ts
@@ -1,10 +1,13 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { rest } from 'msw';
 import { OAuthResponse, RedirectAuthTypes } from '@/api/sign';
-import { USER_LIST as OAUTH_USER_LIST } from '@/components/Molecules/Dropdown/mock';
 import { UserTypes } from '@/api/issue/types';
-import AppLogo from '@/assets/logo/issueTracker.png';
 import { ERROR_CODE } from '@/api/constants';
+
+import { USER_LIST as OAUTH_USER_LIST } from '@/components/Molecules/Dropdown/mock';
+import AppLogo from '@/assets/logo/issueTracker.png';
+
+import { filterIdPassword, getCookie, deleteAllCookies } from '@/mocks/tables/authHelpers';
 import { ACCESS_TOKEN, REFRESH_TOKEN } from '@/mocks/constants';
 
 interface GeneralUserInfoTypes {
@@ -12,7 +15,7 @@ interface GeneralUserInfoTypes {
   password?: string;
 }
 
-type UserTableTypes = GeneralUserInfoTypes & UserTypes;
+export type UserTableTypes = GeneralUserInfoTypes & UserTypes;
 
 export const USER_TABLE: UserTableTypes[] = [
   ...OAUTH_USER_LIST,
@@ -26,45 +29,7 @@ export const USER_TABLE: UserTableTypes[] = [
   },
 ];
 
-const getCookie = (key: string, value: string) => {
-  const matchKey = document.cookie.match(`(^|;) ?${key}=([^;]*)(;|$)`);
-  return matchKey?.filter((el) => el === value)[0];
-};
-
-const deleteAllCookies = () => {
-  const cookies = document.cookie.split('; ');
-  const expiration = 'Fri, 31 Dec 1999 00:00:00 GMT';
-
-  cookies.forEach((cookie: any) => {
-    const [key, _] = cookie.split('=');
-    document.cookie = `${key}=; expires=${expiration}`;
-    document.cookie = `${key}=; expires=${expiration}`;
-    document.cookie = `${key}=; expires=${expiration}; domain=localhost`;
-  });
-};
-
-export const filterIdPassword = (obj: UserTableTypes) =>
-  Object.fromEntries(
-    Object.entries(obj).filter(([key]) => !key.includes('loginId') && !key.includes('password')),
-  ) as UserTypes;
-
 export const TEST_USER = USER_TABLE.find((user) => user.nickname === 'WebTest') || USER_TABLE[0];
-
-export const getAuther = () => {
-  const getSessionStorgeUserInfo = sessionStorage.getItem('user');
-  if (getSessionStorgeUserInfo) {
-    const parserUserInfo = JSON.parse(getSessionStorgeUserInfo) as UserTableTypes;
-    const findUser = USER_TABLE.find((user) => user.nickname === parserUserInfo.nickname);
-
-    if (!findUser) {
-      USER_TABLE.push(parserUserInfo);
-      return filterIdPassword(parserUserInfo);
-    }
-    return filterIdPassword(findUser);
-  }
-
-  return filterIdPassword(TEST_USER);
-};
 
 export const authHandlers = [
   // silent-refresh
